+++ Gemfile
@@ -78,3 +78,4 @@ gem 'font-awesome-sass'
 gem 'carrierwave'
 gem 'jquery-rails'
 gem 'devise'
+gem 'fog-aws'
